Update PrimeCheck.py
I don't understand why range has got third argument since we can accomplish this functionality with just two arguments (we're taking the next value for range with step size 2 and however decrementing it later [in line 5]). So I think, changing it won't change the functionality of the algorithm. Hence I removed 4,5 lines from the code and modified the range iterator to give numbers from 2 to half of the number with step size 1