Refactor
@Lownish Thanks for this submission!

I do not know scheduling and I am struggling to understand this algorithm.  Sometimes it goes in an infinite loop and other times it delivers all zeros.

Can you please provide a set of input data and expected output data?

I created a separate `priority_schedule()` function with a doctest that never seems to calculate a schedule.  Can you please fix that function so that it provides the correct schedule for that input data?