Create a Job class
I hope these changes help to clarify...

I have rearranged the code in a way that will BREAK THE TESTS.   I make this clearer, we will have a Job class and will remove the Schedule class and replace it with functions.  Let's keep the Job class super simple.  Its only goal is to carry around the deadline and profit for each Job so we do not move around separate list of ints.

With these changes, each of our three main functions takes in a list of jobs.  We should keep things flexible so that we could create `test_jobs_two` and run it thru all functions and get different results.