Highest Value Palindrome
Palindromes are strings that read the same from the left or right, for example, madam or 0110.

You will be given a string representation of a number and a maximum number of changes you can make. Alter the string, one digit at a time, to create the string representation of the largest number possible given the limit to the number of changes. The length of the string may not be altered, so you must consider 0 's left of all higher digits in your tests. For example, 0110 is valid,  0011is not.

Given a string representing the starting number and a maximum number of changes allowed, create the largest palindromic string of digits possible or the string -1 if it's impossible to create a palindrome under the constraints.