feat: add algo to group odd even nodes in a Linked List
Adding the algorithm I used to solve
https://leetcode.com/problems/odd-even-linked-list/

This groups odd indexed nodes in a Linked List at the beginning of the
list and then the even indexed nodes following the odd indexed nodes.

Note that the nodes are not grouped by their value but by their
index. This algo runs in O(1) space and O(n) time.

Add geeksforgeeks url for additional details on algorithm