Updated __str__ method for the stack.
Calling print on stack returns Stack object, the new __str__ method returns str object with proper notation of how every element is placed in the stack. Calling print on a stack object now shows proper positioning of the stack elements, also denoting top of the stack and bottom of the stack in proper order.