Update area.py
Updated area.py with code for the area of a triangle when 3 sides are given. The existing area_triangle() function was renamed as area_isosceles_traingle(), as the formula 0.5*base*height is applicable only for isosceles triangles. The general formula for calculating area with 3 sides known is used in area_triangle.