dpll: add example case for recursion bug
The formula {{A3' , A1} , {A3} , {A4}} is satisfiable with the
assignment {'A1': True, 'A4': True, 'A3': True}.

The formula {{A3' , A1} , {A4} , {A3}} is raises a RecursionError and is
not satisfiable.

The formula {{A2' , A3} , {A2 , A3 , A5' , A1} , {A3} , {A5 , A1'}} is
raises a RecursionError and is
not satisfiable.