Prove Fermat's Last Theorem.
Fermat's Last Theorem states that there are no natural integers ($x, y, z, n$) that allow the following condition to be true.

$$
\begin{align}
x^n + y^n = z^n
\end{align}
$$

This theorem is known to be difficult to prove, especially by hand or with a calculator. But if we decide on a maximum test value, high enough that numerical dispersion intuits no higher plausible solution (ex. 4,000, used in the script), we can brute force through all combinations of $x$, $y$, $z$, and $n$ (in about a minute on my computer) to show that no combinations passed the the "solved" test.