Create LittleProfessor.py
Prompts the user for a level, n. If the user does not input 1, 2, or 3, the program should prompt again.
Randomly generates ten (10) math problems formatted as 'X + Y = ', wherein each of X and Y is a non-negative integer with n digits. Only supports addition (+).
Prompts the user to solve each of those problems. If an answer is not correct (or not even a number), the program should output EEE and prompt the user again, allowing the user up to three tries in total. If the user has still not answered correctly after three tries, the program should output the correct answer.
The program ultimately outputs the user’s score, a number out of 10.