tower_of _hanoi.py
Fixes #9936
This code prompts the user to input the number of disks and the names of the source, auxiliary, and target pegs. It also checks that the peg names are distinct. Then, it calls the tower_of_hanoi function to solve the Tower of Hanoi problem based on the user's input.