Create bogosort.py
Bogo sort is a highly inefficient sorting algorithm that is based on the generate and test paradigm. The algorithm successively generates permutations of its input until it finds one that is sorted. It’s not recommended to use this algorithm in practice as it has an average-case time complexity of O(n*n!)